split vm::builtins::iter utility functions to vm::iterator
