+++ vm/src/builtins/filter.rs
@@ -1,6 +1,6 @@
-use super::iter;
 use super::pybool;
 use super::pytype::PyTypeRef;
+use crate::iterator;
 use crate::pyobject::{PyClassImpl, PyContext, PyObjectRef, PyRef, PyResult, PyValue};
 use crate::vm::VirtualMachine;
 
@@ -30,7 +30,7 @@ impl PyFilter {
         iterable: PyObjectRef,
         vm: &VirtualMachine,
     ) -> PyResult<PyRef<Self>> {
-        let iterator = iter::get_iter(vm, &iterable)?;
+        let iterator = iterator::get_iter(vm, &iterable)?;
 
         Self {
             predicate: function,
@@ -44,7 +44,7 @@ impl PyFilter {
         let predicate = &self.predicate;
         let iterator = &self.iterator;
         loop {
-            let next_obj = iter::call_next(vm, iterator)?;
+            let next_obj = iterator::call_next(vm, iterator)?;
             let predicate_value = if vm.is_none(predicate) {
                 next_obj.clone()
             } else {
