+++ vm/src/stdlib/operator.rs
@@ -1,7 +1,7 @@
-use crate::builtins::iter;
 use crate::builtins::pystr::PyStrRef;
 use crate::byteslike::PyBytesLike;
 use crate::function::OptionalArg;
+use crate::iterator;
 use crate::pyobject::{BorrowValue, Either, PyObjectRef, PyResult, TypeProtocol};
 use crate::VirtualMachine;
 use volatile::Volatile;
@@ -14,7 +14,7 @@ fn _operator_length_hint(obj: PyObjectRef, default: OptionalArg, vm: &VirtualMac
             default.class().name
         )));
     }
-    let hint = iter::length_hint(vm, obj)?
+    let hint = iterator::length_hint(vm, obj)?
         .map(|i| vm.ctx.new_int(i))
         .unwrap_or(default);
     Ok(hint)
